// whirlpool/whirlpoolemea/cc-google-tag-manager/react/typings/ProductTypes.ts
export type ProductSpecification = {
  name: string;
  originalName: string;
  values: string[];
};

/** @deprecated Use `ProductSpecification` */
export type SpecificationGroupItem = ProductSpecification;

export type SpecificationGroup = {
  name: string;
  originalName: string;
  specifications: ProductSpecification[];
};

export type SecondaryRatingsAverages = {
  Id: string;
  AverageRating: number;
}